Extract common revision diff context menu logic into IRevisionDiffContextMenuController
